+++ ashley/src/main/kotlin/ktx/ashley/dsl.kt
@@ -1,0 +1,8 @@
+package ktx.ashley
+
+/**
+ * Marks KTX Ashley type-safe builders.
+ */
+@DslMarker
+@Target(AnnotationTarget.CLASS, AnnotationTarget.TYPE)
+annotation class AshleyDsl